[doc] Extract template-fetch config from conf.py into its own module
Move the build-time sphinx-collections template-fetch machinery (network
constants + retry, the template registry, _resolve/_fetch helpers, the
collections config, the _collections/ exclude_patterns and ipython3 lexer
pattern, and the _collections/ footnote filter + code-block retag hooks)
out of conf.py into doc/source/template_collections.py. conf.py imports
the config values and calls template_collections.register(app) in setup.

Pure move, no behavior change: _TEMPLATE_COLLECTIONS is byte-identical,
exclude_patterns() reproduces the same _collections entries, and importing
conf.py resolves the integration. This scopes template-publishing changes
to one module (for CI/ownership and the upcoming version pinning) and keeps
conf.py focused on Sphinx configuration.
